feat: update .clang-tidy checks and add NOLINT comments for readabili…

+++ src/data_structures/ast.c
@@ -4,6 +4,7 @@
 #include <stdlib.h>
 
 #include "core.h"
+#include "data_structures/container.h"
 #include "debug.h"
 #include "syntax.h"
 
@@ -92,7 +93,7 @@ void print_ast_node(ASTNode* node, const char* prefix, bool is_last) {
         node_name = "UNKNOWN_NODE";
     }
 
-    const char* marker = is_last ? "└── " : "├── ";
+    const char* marker = (int)is_last ? "└── " : "├── ";
     if (fprintf(logs.ast, "%s%s%s", prefix, marker, node_name) < 0) {
         jakarta_error(ERR_CUSTOM, nullptr, "Failed to print AST node");
     }
@@ -109,7 +110,7 @@ void print_ast_node(ASTNode* node, const char* prefix, bool is_last) {
 
     char new_prefix[AST_PRINT_BUFFER_SIZE];
     if (snprintf(new_prefix, sizeof(new_prefix), "%s%s", prefix,
-                 is_last ? "    " : "│   ") < 0) {
+                 (int)is_last ? "    " : "│   ") < 0) {
         jakarta_error(ERR_CUSTOM, nullptr, "Failed to create new prefix");
     }
